ssl_cert_provider: make BIOMem class external
